Fix compilation error for minimad example

+++ minimad.c
@@ -176,7 +176,7 @@ enum mad_flow error(void *data,
 
   fprintf(stderr, "decoding error 0x%04x (%s) at byte offset %u\n",
 	  stream->error, mad_stream_errorstr(stream),
-	  stream->this_frame - buffer->start);
+	  (unsigned int)(stream->this_frame - buffer->start));
 
   /* return MAD_FLOW_BREAK here to stop decoding (and propagate an error) */
